All types of ads completely implemented. Added service for Context resolution
Added all the remaining ads implementations.  Added the ContextResolver service in the extensions for Dependency Injection.

+++ sources/dotnet/AMDevIT.Admob.Wrapper/AMDevIT.Admob.Wrapper.MAUICross/AdReward.cs
@@ -1,12 +1,12 @@
 ﻿namespace AMDevIT.Admob.Wrapper.MAUICross;
 
-public class AdReward(string type, int amount)
+public class AdReward(string type, long amount)
 {
     #region Properties
 
     public string Type => type;
 
-    public int Amount => amount;
+    public long Amount => amount;
 
     #endregion
